(FM-8117) Add Windows Support
Prior to this commit the only supported client platform for
using KREAM was macos/linux. This commit adds a PowerShell
psake build script with rough parity for the rake file.

It also updates the Vagrantfile to default to Hyper-V if run
on Windows.

There are a few drawbacks to this approach on windows:

- Due to limitations with Hyper-V and Vagrant, the spinning
  up of vagrant machines is in serial in the foreground as
  you need to manually specify the network stack and the
  password for the shared drive.
- There is no logging in place as the actions are in the
  foreground and written to the console.