(maint) - add support for kubernetes v1.25+
Prior to this commit, kream was unable to spin up clusters with
kubernetes v1.25 or above.
This was due to the deprecation of certain containerd versions and the
updating of the link to kubernetes container registry.

This commit aims to add two hiera variables
('kubernetes::containerd_version', 'kubernetes::image_repository')
that can be adjusted so that kream can spin up clusters with
kubernetes versions that are both above and below v1.25.